+++ src/merge.rs
@@ -1,4 +1,4 @@
-// Copyright 2018, 2019, 2020 Martin Pool.
+// Copyright 2018, 2019, 2020, 2021 Martin Pool.
 
 // This program is free software; you can redistribute it and/or modify
 // it under the terms of the GNU General Public License as published by
@@ -20,21 +20,26 @@ use std::cmp::Ordering;
 use crate::*;
 
 #[derive(Debug, PartialEq, Eq)]
-pub enum MergedEntryKind {
-    LeftOnly,
-    RightOnly,
-    Both,
-    // TODO: Perhaps also include the tree-specific entry kind?
+pub enum MergedEntryKind<AE, BE>
+where
+    AE: Entry,
+    BE: Entry,
+{
+    LeftOnly(AE),
+    RightOnly(BE),
+    Both(AE, BE),
 }
 
 use self::MergedEntryKind::*;
 
 #[derive(Debug, PartialEq, Eq)]
-pub struct MergedEntry {
-    // TODO: Add accessors rather than making these public?
-    // TODO: Include the original entries from either side?
+pub struct MergedEntry<AE, BE>
+where
+    AE: Entry,
+    BE: Entry,
+{
     pub apath: Apath,
-    pub kind: MergedEntryKind,
+    pub kind: MergedEntryKind<AE, BE>,
 }
 
 /// Zip together entries from two trees, into an iterator of MergedEntryKind.
@@ -77,7 +82,7 @@ where
     AE: Entry,
     BE: Entry,
 {
-    type Item = MergedEntry;
+    type Item = MergedEntry<AE, BE>;
 
     fn next(&mut self) -> Option<Self::Item> {
         // TODO: Stats about the merge.
@@ -86,8 +91,7 @@ where
         // Preload next-A and next-B, if they're not already
         // loaded.
         //
-        // TODO: Perhaps use <https://doc.rust-lang.org/stable/core/iter/struct.Peekable.html> instead of keeping a
-        // readahead here?
+        // TODO: Perhaps use `Peekable` instead of keeping a readahead here?
         if self.na.is_none() {
             self.na = ait.next();
         }
@@ -101,40 +105,31 @@ where
                 let tb = self.nb.take().unwrap();
                 Some(MergedEntry {
                     apath: tb.apath().clone(),
-                    kind: RightOnly,
+                    kind: RightOnly(tb),
                 })
             }
         } else if self.nb.is_none() {
+            let ta = self.na.take().unwrap();
             Some(MergedEntry {
-                apath: self.na.take().unwrap().apath().clone(),
-                kind: LeftOnly,
+                apath: ta.apath().clone(),
+                kind: LeftOnly(ta),
             })
         } else {
             let pa = self.na.as_ref().unwrap().apath().clone();
             let pb = self.nb.as_ref().unwrap().apath().clone();
             match pa.cmp(&pb) {
-                Ordering::Equal => {
-                    self.na.take();
-                    self.nb.take();
-                    Some(MergedEntry {
-                        apath: pa,
-                        kind: Both,
-                    })
-                }
-                Ordering::Less => {
-                    self.na.take().unwrap();
-                    Some(MergedEntry {
-                        apath: pa,
-                        kind: LeftOnly,
-                    })
-                }
-                Ordering::Greater => {
-                    self.nb.take().unwrap();
-                    Some(MergedEntry {
-                        apath: pb,
-                        kind: RightOnly,
-                    })
-                }
+                Ordering::Equal => Some(MergedEntry {
+                    apath: pa,
+                    kind: Both(self.na.take().unwrap(), self.nb.take().unwrap()),
+                }),
+                Ordering::Less => Some(MergedEntry {
+                    apath: pa,
+                    kind: LeftOnly(self.na.take().unwrap()),
+                }),
+                Ordering::Greater => Some(MergedEntry {
+                    apath: pb,
+                    kind: RightOnly(self.nb.take().unwrap()),
+                }),
             }
         }
     }
@@ -142,7 +137,6 @@ where
 
 #[cfg(test)]
 mod tests {
-    use super::MergedEntry;
     use super::MergedEntryKind::*;
     use crate::test_fixtures::*;
     use crate::*;
@@ -157,13 +151,16 @@ mod tests {
         )
         .collect::<Vec<_>>();
         assert_eq!(di.len(), 1);
-        assert_eq!(
-            di[0],
-            MergedEntry {
-                apath: "/".into(),
-                kind: Both,
+        assert_eq!(di[0].apath, "/");
+        match &di[0].kind {
+            Both(ae, be) => {
+                assert_eq!(ae.kind(), Kind::Dir);
+                assert_eq!(be.kind(), Kind::Dir);
+                assert_eq!(ae.apath(), "/");
+                assert_eq!(be.apath(), "/");
             }
-        );
+            other => panic!("unexpected {:#?}", other),
+        }
     }
 
     // TODO: More tests of various diff situations.